Infra for worker cleanup
